+++ TGA_FTIR_tools/plotting/__init__.py
@@ -8,6 +8,7 @@ from .plot_mass_step import plot_mass_steps
 from .plot_results import bar_plot_results
 from .plot_rob import plot_robustness
 from .plots import plots
+from .plot_cali import plot_integration, plot_calibration_single, plot_calibration_combined, plot_residuals_single
 from .plotting import FTIR_to_DTG, plot_FTIR, plot_TGA
 from .utils import get_label, make_title